ai-agent/requirements.txt
fastapi
uvicorn
langfuse
requests
python-dotenv
psycopg2-binary
pandas
openpyxl
